+++ tests/snapshot_tests/test_snapshots.py
@@ -945,6 +945,11 @@ def hello(name):
     )
 
 
+def test_text_area_alternate_screen(snap_compare):
+    assert snap_compare(
+        SNAPSHOT_APPS_DIR / "text_area_alternate_screen.py", terminal_size=(48, 10)
+    )
+
 @pytest.mark.syntax
 def test_text_area_wrapping_and_folding(snap_compare):
     assert snap_compare(
@@ -1071,6 +1076,16 @@ def test_recompose(snap_compare):
     """Check recompose works."""
     # https://github.com/Textualize/textual/pull/4206
     assert snap_compare(SNAPSHOT_APPS_DIR / "recompose.py")
+
+@pytest.mark.parametrize("dark", [True, False])
+def test_ansi_color_mapping(snap_compare, dark):
+    """Test how ANSI colors in Rich renderables are mapped to hex colors."""
+
+    def setup(pilot):
+        pilot.app.dark = dark
+
+    assert snap_compare(SNAPSHOT_APPS_DIR / "ansi_mapping.py", run_before=setup)
+
 
 def test_pretty_grid_gutter_interaction(snap_compare):
     """Regression test for https://github.com/Textualize/textual/pull/4219."""